arun commit 4

+++ src/main/java/com/capgemini/hcm/entity/Users.java
@@ -1,5 +1,7 @@
 package com.capgemini.hcm.entity;
 
+import java.math.BigInteger;
+
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
@@ -34,7 +36,7 @@ public class Users {
 
 	@NotNull(message = "contact must be mandatory")
 	@Column(name = "contactNo")
-	private Integer contactNo;
+	private BigInteger contactNo;
 
 	@Column(name = "userRole")
 	private String userRole;
@@ -67,11 +69,11 @@ public class Users {
 		this.userName = userName;
 	}
 
-	public Integer getContactNo() {
+	public BigInteger getContactNo() {
 		return contactNo;
 	}
 
-	public void setContactNo(Integer contactNo) {
+	public void setContactNo(BigInteger contactNo) {
 		this.contactNo = contactNo;
 	}
 
@@ -94,7 +96,7 @@ public class Users {
 	public Users(Integer userId,
 			@NotEmpty(message = "user password is mandatory") @Pattern(regexp = "^(?=.*?[A-Z])(?=.*?[a-z])(?=.*?[0-9])(?=.*?[#?!@$%^&*-]).{4,14}$") String userPassword,
 			@NotEmpty(message = "user name is mandatory") String userName,
-			@NotNull(message = "contact must be mandatory") @Size(max = 10, message = "number must be of 10 digit") Integer contactNo,
+			@NotNull(message = "contact must be mandatory") @Size(max = 10, message = "number must be of 10 digit") BigInteger contactNo,
 			String userRole, @NotEmpty(message = "emailid is mandatory") String emailId) {
 		super();
 		this.userId = userId;
